+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Pending
+
+### Added
+
+- Add Dramatiq integration
+  ([PR #223](https://github.com/scoutapp/scout_apm_python/pull/223)).
+
 ## [2.3.0] 2019-08-04
 
 ### Added